backend/requirements.txt
django
djangorestframework
mysqlclient


# commands - python manage.py makemigrations || python manage.py migrate